Moved tlv utils functions to their own file
