+++ src/features/generic_tx_parser/gtp_param_enum.c
@@ -6,7 +6,7 @@
 #include "shared_context.h"
 #include "tx_ctx.h"
 #include "tlv_library.h"
-#include "tlv_apdu.h"
+#include "tlv_utils.h"
 
 #define PARAM_ENUM_TAGS(X)                                   \
     X(0x00, TAG_VERSION, handle_version, ENFORCE_UNIQUE_TAG) \
