+++ src/features/generic_tx_parser/gtp_param_network.c
@@ -16,7 +16,7 @@
 #include "gtp_field_table.h"
 #include "tx_ctx.h"
 #include "apdu_constants.h"
-#include "tlv_apdu.h"
+#include "tlv_utils.h"
 #include "chain_config.h"
 
 /**
