+++ src/features/generic_tx_parser/gtp_param_nft.c
@@ -3,7 +3,7 @@
 #include "utils.h"
 #include "gtp_field_table.h"
 #include "tlv_library.h"
-#include "tlv_apdu.h"
+#include "tlv_utils.h"
 
 #define PARAM_NFT_TAGS(X)                                    \
     X(0x00, TAG_VERSION, handle_version, ENFORCE_UNIQUE_TAG) \
