+++ src/features/generic_tx_parser/gtp_param_raw.c
@@ -8,7 +8,7 @@
 #include "utils.h"
 #include "shared_context.h"
 #include "tlv_library.h"
-#include "tlv_apdu.h"
+#include "tlv_utils.h"
 
 #define PARAM_RAW_TAGS(X)                                    \
     X(0x00, TAG_VERSION, handle_version, ENFORCE_UNIQUE_TAG) \
