+++ src/features/generic_tx_parser/gtp_param_token.c
@@ -7,7 +7,7 @@
 #include "gtp_field_table.h"
 #include "tx_ctx.h"
 #include "tlv_library.h"
-#include "tlv_apdu.h"
+#include "tlv_utils.h"
 
 #define PARAM_TOKEN_TAGS(X)                                  \
     X(0x00, TAG_VERSION, handle_version, ENFORCE_UNIQUE_TAG) \
