+++ src/features/generic_tx_parser/gtp_param_unit.c
@@ -2,7 +2,7 @@
 #include "gtp_field_table.h"
 #include "shared_context.h"
 #include "tlv_library.h"
-#include "tlv_apdu.h"
+#include "tlv_utils.h"
 
 #define PARAM_UNIT_TAGS(X)                                     \
     X(0x00, TAG_VERSION, handle_version, ENFORCE_UNIQUE_TAG)   \
