+++ src/features/generic_tx_parser/gtp_tx_info.c
@@ -10,7 +10,7 @@
 #include "app_mem_utils.h"
 #include "tx_ctx.h"
 #include "tlv_library.h"
-#include "tlv_apdu.h"
+#include "tlv_utils.h"
 
 #define TX_INFO_TAGS(X)                                                            \
     X(0x00, TAG_VERSION, handle_version, ENFORCE_UNIQUE_TAG)                       \
