+++ src/features/generic_tx_parser/gtp_value.c
@@ -8,7 +8,7 @@
 #include "ui_utils.h"
 #include "tx_ctx.h"
 #include "tlv_library.h"
-#include "tlv_apdu.h"
+#include "tlv_utils.h"
 
 #define VALUE_TAGS(X)                                                      \
     X(0x00, TAG_VERSION, handle_version, ENFORCE_UNIQUE_TAG)               \
