+++ src/features/provide_safe_account/safe_descriptor.c
@@ -6,7 +6,7 @@
 #include "public_keys.h"
 #include "lcx_ecdsa.h"
 #include "tlv_library.h"
-#include "tlv_apdu.h"
+#include "tlv_utils.h"
 #include "utils.h"
 
 #define TYPE_LESM_ACCOUNT_INFO 0x27
