+++ src/features/provide_tx_simulation/cmd_get_tx_simulation.c
@@ -7,6 +7,7 @@
 #include "get_public_key.h"
 #include "tlv_library.h"
 #include "tlv_apdu.h"
+#include "tlv_utils.h"
 #include "utils.h"
 #include "nbgl_use_case.h"
 #include "os_pki.h"
