+++ src/features/sign_authorization_eip7702/auth_7702.c
@@ -1,6 +1,6 @@
 #include "auth_7702.h"
 #include "utils.h"
-#include "tlv_apdu.h"
+#include "tlv_utils.h"
 
 #define STRUCT_VERSION 0x01
